ADD scrpits for create latebreaking abstracts data
Late breaking abstract data parsed to format needed for abstract book
Rename the data from the first batch of submission

+++ ohbm2021_abstract_book/parser_web.py
@@ -61,22 +61,39 @@ def web2list():
     return o["rst"]["abstracts"]["abstract"]
 
 
-def main():
-    abstract_list = web2list()
-    consolidated_dict = {}
-    print(len(abstract_list))
-    for original in abstract_list:
-        current = original.copy()
-        instance_dict, _ = compile_pd_dict(current, {})
-        if original.get("@id") not in consolidated_dict:
-            consolidated_dict.update({original.get("@id"): instance_dict.copy()})
-        else:
-            print("ID exist...")
-    print(len(consolidated_dict))
-    df = pd.DataFrame(consolidated_dict).T
-    df = df[df_cols]
-    df.to_pickle("abstracts_content.pkl")
+def read_latebreaking(poster_id_file="latebreaking_abstracts_list.csv"):
+    with open(poster_id_file, "r") as f:
+        abstract_id = [item.split()[0]
+                       for item in f.readlines()]
+    return abstract_id
 
+
+def main(late_break=False):
+    if late_break and os.path.isfile("abstracts_content.pkl"):
+        df = pd.read_pickle("abstracts_content.pkl")
+        late_abstract_list = read_latebreaking()
+        print(len(late_abstract_list))
+        bool_late = df['abstractNumber'].isin(late_abstract_list)
+        print(sum(bool_late))
+        df_late = df[bool_late]
+        df_late.to_pickle("latebreak_abstracts_content.pkl")
+        return df_late
+    else:
+        abstract_list = web2list()
+        consolidated_dict = {}
+        print(len(abstract_list))
+        for original in abstract_list:
+            current = original.copy()
+            instance_dict, _ = compile_pd_dict(current, {})
+            if original.get("@id") not in consolidated_dict:
+                consolidated_dict.update({original.get("@id"): instance_dict.copy()})
+            else:
+                print("ID exist...")
+        print(len(consolidated_dict))
+        df = pd.DataFrame(consolidated_dict).T
+        df = df[df_cols]
+        df.to_pickle("abstracts_content.pkl")
+        return df
 
 if __name__ == "__main__":
     main()
